requirements.txt
deepface==0.0.93
faiss-cpu
numpy~=2.0.0
pandas
pillow~=11.1.0
pydantic
PyYAML
seaborn
tf-keras
scikit-learn~=1.6.0
sqlalchemy~=2.0.0
